refactor(ui): promote the credential + permission-rule forms into shared/
Relocates the credential form closure and the permission-rule editor out of
their feature modules into shared/ so any surface can compose them — the
prerequisite for the provisioning-plan fulfilment wizard, which must reuse both
the credential form (credentials module) and the rule editor (toolkits module)
from a single place. ESLint forbids shared/ importing modules/ and forbids
sibling-module imports, so these had to move to be co-composable.

No behavior change (pure move + import rewiring; 691/691 UI tests pass):
- src/modules/credentials/{api,components,lib,config.ts,mocks} + their unit
  tests -> src/shared/credentials/ (pages/ and routes.tsx stay in the module).
- PermissionRuleEditor (+ PERMISSION_EFFECTS/PermissionEffect/PermissionRuleInput)
  -> src/shared/ui/PermissionRuleEditor.tsx, exported from the shared/ui barrel;
  toolkits/api/types re-exports the moved types so its public surface is stable.
- rewired all @/modules/credentials imports to @/shared/credentials; module
  routes/pages/mocks references preserved.
