+++ ui/src/shared/credentials/__tests__/scope-utils.test.ts
@@ -9,7 +9,7 @@ import {
 	isRecommendedScope,
 	scopesInGroup,
 	type EnhancedScope,
-} from '@/modules/credentials/lib/scope-utils';
+} from '@/shared/credentials/lib/scope-utils';
 
 function scope(name: string, recommended = false, description = ''): EnhancedScope {
 	return { scope: name, description, origin: 'schema', isRecommended: recommended };
