+++ ui/src/shared/credentials/components/ServerVariablesSection.tsx
@@ -1,6 +1,6 @@
 import { useId } from 'react';
 import { Input, Label, Select } from '@/shared/ui';
-import type { ServerVarDef } from '@/modules/credentials/api';
+import type { ServerVarDef } from '@/shared/credentials/api';
 
 /**
  * Renders an input per OpenAPI **server variable** (e.g. Atlassian's
